refactor: colocando cada page com seu css
