+++ src/frontend/src/pages/Home/Home.css
@@ -1,0 +1,42 @@
+/* ===== Home Page ===== */
+.page-full {
+  min-height: 100vh;
+}
+
+.main-content {
+  padding-top: 2rem;
+  padding-bottom: 2rem;
+  display: flex;
+  flex-direction: column;
+  gap: 1.5rem;
+}
+
+/* ===== Support Card ===== */
+.support-info {
+  background: var(--gradient-soft);
+  border-radius: 0.75rem;
+  padding: 1rem;
+  display: flex;
+  flex-direction: column;
+  gap: 0.25rem;
+}
+
+.support-name {
+  font-size: 0.875rem;
+  font-weight: 600;
+  color: var(--foreground);
+}
+
+.support-phone {
+  font-size: 0.875rem;
+  font-weight: 500;
+  color: var(--primary);
+  display: flex;
+  align-items: center;
+  gap: 0.375rem;
+}
+
+.support-phone svg {
+  width: 0.875rem;
+  height: 0.875rem;
+}
