+++ src/frontend/src/pages/Home/Home.tsx
@@ -1,3 +1,4 @@
+import "./Home.css";
 import { motion } from "framer-motion";
 import Navbar from "@/components/Navbar";
 import MoodSelector from "@/components/MoodSelector";
