+++ src/frontend/src/pages/NotFound/NotFound.tsx
@@ -1,3 +1,4 @@
+import "./NotFound.css";
 import { useLocation } from "react-router-dom";
 import { useEffect } from "react";
